requirements.txt
streamlit>=1.0
python-dotenv>=0.21
mail-parser>=3.0
openai>=0.27
pyairtable>=1.0
